+++ escalation/tests/test_wizard_utils.py
@@ -27,6 +27,7 @@ from utility.constants import (
     SELECTABLE_DATA_DICT,
     VISUALIZATION_OPTIONS,
     GROUPBY_SELECTOR,
+    MAX_ENTRIES_FOR_FILTER_SELECTOR,
 )
 from utility.wizard_utils import (
     invert_dict_lists,
@@ -39,7 +40,10 @@ from utility.wizard_utils import (
     graphic_dict_to_graphic_component_dict,
     graphic_component_dict_to_graphic_dict,
     generate_collapse_dict_from_graphic_component_dict,
+    divide_columns_into_type_of_filters,
 )
+
+from sqlalchemy.types import Integer, Text, Float, DateTime, ARRAY, Boolean
 
 
 def test_invert_dict_lists():
@@ -49,7 +53,7 @@ def test_invert_dict_lists():
     assert inverted_dict == expected_dict
 
 
-def test_get_layout_for_dashboard(main_json_csv_backend_fixture):
+def test_get_layout_for_dashboard(main_json_sql_backend_fixture):
     # Need to set up file system
     # new_dict = get_layout_for_dashboard(main_json_csv_backend_fixture[AVAILABLE_PAGES])
     #
@@ -143,8 +147,8 @@ def test_extract_data_sources_from_config(graphic_json_fixture):
     assert len(data_sources) == 2
 
 
-def test_copy_data_from_form_to_config(main_json_csv_backend_fixture):
-    main_config = copy.deepcopy(main_json_csv_backend_fixture)
+def test_copy_data_from_form_to_config(main_json_sql_backend_fixture):
+    main_config = copy.deepcopy(main_json_sql_backend_fixture)
     placer_holder_text = (
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod"
         " tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam"
@@ -166,8 +170,15 @@ def test_copy_data_from_form_to_config(main_json_csv_backend_fixture):
 def test_graphic_dict_to_graphic_component_dict(graphic_json_fixture):
     test_config = graphic_json_fixture["graphic_1"]
     component_dict = graphic_dict_to_graphic_component_dict(test_config)
-    assert component_dict[VISUALIZATION] == {}
+    assert component_dict[VISUALIZATION] == {
+        "aggregate": {},
+        "groupby": {},
+        "hover_data": {},
+    }
     assert component_dict[PLOTLY] == test_config[PLOT_SPECIFIC_INFO]
+    selector_dict_expected = test_config[SELECTABLE_DATA_DICT]
+    selector_dict_expected[NUMERICAL_FILTER] = []
+    selector_dict_expected[FILTER] = []
     assert component_dict[SELECTOR] == test_config[SELECTABLE_DATA_DICT]
     test_config_copy = copy.deepcopy(test_config)
     # component_dict[GRAPHIC_META_INFO] contains all the data besides the keys
@@ -175,6 +186,7 @@ def test_graphic_dict_to_graphic_component_dict(graphic_json_fixture):
     # test_config_copy does not contain VISUALIZATION_OPTIONS and we delete the other two.
     del test_config_copy[PLOT_SPECIFIC_INFO]
     del test_config_copy[SELECTABLE_DATA_DICT]
+    test_config_copy[DATA_SOURCES][ADDITIONAL_DATA_SOURCES] = []
     assert component_dict[GRAPHIC_META_INFO] == test_config_copy
 
 
@@ -205,3 +217,56 @@ def test_generate_collapse_dict_from_graphic_component_dict(graphic_json_fixture
         SELECTABLE_DATA_DICT: False,
     }
     assert collapse_dict == true_collapse_dict
+
+
+def test_divide_columns_into_type_of_filters():
+    column_types_dict = {
+        "col_1": Integer(),
+        "col_2": Text(),
+        "col_3": Float(),
+        "col_4": DateTime(),
+        "col_5": ARRAY("string"),
+        "col_6": Boolean(),
+        "col_7": Integer(),
+        "col_8": Float(),
+        "col_9": DateTime(),
+        "col_10": Text(),
+        "col_11": Boolean(),
+        "col_12": ARRAY("string"),
+    }
+    unique_entries = {
+        "col_1": [1 for i in range(MAX_ENTRIES_FOR_FILTER_SELECTOR + 1)],
+        "col_2": ["Dream" for i in range(MAX_ENTRIES_FOR_FILTER_SELECTOR - 1)],
+        "col_3": [2.5 for i in range(MAX_ENTRIES_FOR_FILTER_SELECTOR + 1)],
+        "col_4": ["11/10/2013" for i in range(MAX_ENTRIES_FOR_FILTER_SELECTOR + 1)],
+        "col_5": [[1, 4, 5] for i in range(MAX_ENTRIES_FOR_FILTER_SELECTOR - 1)],
+        "col_6": [True for i in range(MAX_ENTRIES_FOR_FILTER_SELECTOR - 1)],
+        "col_7": [1 for i in range(MAX_ENTRIES_FOR_FILTER_SELECTOR - 1)],
+        "col_8": [3.5 for i in range(MAX_ENTRIES_FOR_FILTER_SELECTOR - 1)],
+        "col_9": ["11/10/2014" for i in range(MAX_ENTRIES_FOR_FILTER_SELECTOR - 1)],
+        "col_10": ["Dream" for i in range(MAX_ENTRIES_FOR_FILTER_SELECTOR + 1)],
+        "col_11": [True for i in range(MAX_ENTRIES_FOR_FILTER_SELECTOR + 1)],
+        "col_12": [[1, 4, 5] for i in range(MAX_ENTRIES_FOR_FILTER_SELECTOR + 1)],
+    }
+    (
+        filter_column_names,
+        numerical_filter_column_names,
+        unique_entries_dict,
+    ) = divide_columns_into_type_of_filters(unique_entries, column_types_dict)
+    assert set(numerical_filter_column_names) == {
+        "col_1",
+        "col_3",
+        "col_4",
+        "col_7",
+        "col_8",
+        "col_9",
+    }
+    filter_column_names_set = {"col_2", "col_5", "col_6", "col_7", "col_8", "col_9"}
+    assert set(filter_column_names) == filter_column_names_set
+    assert set(unique_entries_dict.keys()) == filter_column_names_set
+    assert unique_entries_dict["col_2"] == unique_entries["col_2"]
+    assert unique_entries_dict["col_5"] == unique_entries["col_5"]
+    assert unique_entries_dict["col_6"] == unique_entries["col_6"]
+    assert unique_entries_dict["col_7"] == unique_entries["col_7"]
+    assert unique_entries_dict["col_8"] == unique_entries["col_8"]
+    assert unique_entries_dict["col_9"] == unique_entries["col_9"]